Add multiple groupby commands and validation
Add: Validation for the groupby command and return error if value passed is not a valid command

Add: Allow for multiple values for the groupby command
