+++ pkg/cli/cli.go
@@ -3,13 +3,12 @@ package cli
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	"github.com/Boeing/config-file-validator/pkg/finder"
 	"github.com/Boeing/config-file-validator/pkg/reporter"
 )
 
-var GroupOutput string
+var GroupOutput []string
 
 type CLI struct {
 	// FileFinder interface to search for the files
@@ -39,7 +38,7 @@ func WithReporter(reporter reporter.Reporter) CLIOption {
 	}
 }
 
-func WithGroupOutput(groupOutput string) CLIOption {
+func WithGroupOutput(groupOutput []string) CLIOption {
 	return func(c *CLI) {
 		GroupOutput = groupOutput
 	}
@@ -97,17 +96,17 @@ func (c CLI) Run() (int, error) {
 		reports = append(reports, report)
 	}
 
-	switch {
-	case GroupOutput == "filetype":
-		reports = GroupByFile(reports)
-		c.Reporter.Print(reports)
-	case GroupOutput == "pass/fail":
-		reports = GroupByPassFail(reports)
-		c.Reporter.Print(reports)
-	case GroupOutput == "directory":
-		reports = GroupByDirectory(reports)
-		c.Reporter.Print(reports)
-	default:
+	if len(GroupOutput) > 0 {
+		for _, group := range GroupOutput {
+			switch group {
+			case "filetype":
+				reports = GroupByFile(reports)
+			case "pass/fail":
+				reports = GroupByPassFail(reports)
+			case "directory":
+				reports = GroupByDirectory(reports)
+			}
+		}
 		c.Reporter.Print(reports)
 	}
 
@@ -116,73 +115,4 @@ func (c CLI) Run() (int, error) {
 	} else {
 		return 0, nil
 	}
-}
-
-// Group Files by File Type
-func GroupByFile(reports []reporter.Report) []reporter.Report {
-
-    mapFiles := make(map[string][]reporter.Report)
-    reportByFile := []reporter.Report{}
-
-	for _, report := range reports {
-        fileType := strings.Split(report.FileName, ".")[1]
-        if mapFiles[fileType] == nil {
-            mapFiles[fileType] = []reporter.Report{report}
-        } else {
-            mapFiles[fileType] = append(mapFiles[fileType], report)
-        }
-	}
-
-    for _, reports := range mapFiles {
-        reportByFile = append(reportByFile, reports...)
-    }
-
-	return reportByFile
-}
-
-func GroupByPassFail(reports []reporter.Report) []reporter.Report {
-    mapFiles := make(map[string][]reporter.Report)
-    reportByPassOrFail := []reporter.Report{}
-
-	for _, report := range reports {
-        if report.IsValid {
-            if mapFiles["pass"] == nil {
-                mapFiles["pass"] = []reporter.Report{report}
-            } else {
-                mapFiles["pass"] = append(mapFiles["pass"], report)
-            }
-        } else {
-            if mapFiles["fail"] == nil {
-                mapFiles["fail"] = []reporter.Report{report}
-            } else {
-                mapFiles["fail"] = append(mapFiles["fail"], report)
-            }
-        }
-    }
-
-    for _, reports := range mapFiles {
-        reportByPassOrFail = append(reportByPassOrFail, reports...)
-    }
-
-	return reportByPassOrFail
-}
-
-func GroupByDirectory(reports []reporter.Report) []reporter.Report {
-    mapFiles := make(map[string][]reporter.Report)
-    reportByDirectory := []reporter.Report{}
-
-    for _, report := range reports {
-        directory := strings.Split(report.FilePath, "/")[1]
-        if mapFiles[directory] == nil {
-            mapFiles[directory] = []reporter.Report{report}
-        } else {
-            mapFiles[directory] = append(mapFiles[directory], report)
-        }
-    }
-
-    for _, reports := range mapFiles {
-        reportByDirectory = append(reportByDirectory, reports...)
-    }
-
-    return reportByDirectory
 }
